+++ deployments/the-web-cloud-platform-1/pipeline.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env sh
 set -e
 
-helm chart pull "localhost:55000/webdock/the-web-cloud-platform:0.0.0"
-helm chart export "localhost:55000/webdock/the-web-cloud-platform:0.0.0"
+# TODO GitHub packages does not support Helm yet (https://github.community/t5/GitHub-Actions/Feature-Request-Support-Helm-Charts-in-GitHub-Package-Registry/td-p/39186)
+#helm chart pull "docker.pkg.github.com/damlys/webdock/the-web-cloud-platform:0.0.0"
+helm chart export "docker.pkg.github.com/damlys/webdock/the-web-cloud-platform:0.0.0"
 
 helm \
   --kube-context="minikube" \